+++ src/Components/MyTrips.js
@@ -54,29 +54,6 @@ function MyTripsPage()
       )
     );
   }
-
-  //map PastTrips into node
-  // let displayPastTrips = <h3 className="text-center">You haven't gone on any trips... yet :D </h3>;
-  // if (PastTrips.length > 0) {
-  //   displayPastTrips = PastTrips.map(trip => 
-  //     <Col key={trip.name}>
-  //       <div>
-  //         <TripCard {...trip} handleTripsMod={handleTripsMod} date={(trip.ItineraryObject.startDate.toDateString()).slice(4)+' - '+(trip.ItineraryObject.endDate.toDateString()).slice(4)} srcImg={trip.srcImg}></TripCard>
-  //       </div>
-  //     </Col>
-  //     );
-  // }
-  
-  // map TripArray into node
-  // if (TripArray.length > 0) {
-  //   displayUpcomingTrips = TripArray.map(trip => 
-  //     <Col key={trip.name}>
-  //       <div>
-  //         <TripCard {...trip} date={(trip.ItineraryObject.startDate.toDateString()).slice(4)+' - '+(trip.ItineraryObject.endDate.toDateString()).slice(4)} srcImg={trip.srcImg}></TripCard>
-  //       </div>
-  //     </Col>
-  //     );
-  // }
   
   // display on page
   return (
